+++ assets/javascripts/kitten/components/modals/modal/index.js
@@ -87,7 +87,9 @@ export class Modal extends Component {
           ariaHideApp={false}
           onRequestClose={this.close}
           contentLabel={label}
-          bodyOpenClassName={disableOutsideScroll && 'k-Modal__body--open'}
+          bodyOpenClassName={
+            disableOutsideScroll ? 'k-Modal__body--open' : null
+          }
           {...modalProps}
         >
           {content}